+++ services/linkedin.js
@@ -22,4 +22,24 @@ if (Meteor.isClient) {
         var credentialRequestCompleteCallback   =   Accounts.oauth.credentialRequestCompleteHandler(callback);
             LinkedIn.requestCredential(options, credentialRequestCompleteCallback);
     };
+
+    Meteor.getLinkedInInfo  =   function(options, callback) {
+        if ( Meteor.userId() ) {
+            throw new Meteor.Error(401, "You're already logged in.");
+        }
+
+        if ( !callback && (typeof options === "function") ) {
+            callback    =   options;
+            options     =   null;
+        }
+
+        LinkedIn.requestCredential(options, function (credentialTokenOrError) {
+            if ( credentialTokenOrError && (credentialTokenOrError instanceof Error) ) {
+                callback(credentialTokenOrError);
+            } else {
+                var credentialSecret    =   OAuth._retrieveCredentialSecret(credentialTokenOrError);
+                Meteor.call("LinkedIn.__retrieveCredential", credentialTokenOrError, credentialSecret, callback);
+            }
+        });
+    }
 }