+++ history.md
@@ -12,3 +12,15 @@ v 0.0.3
 * google support 
 * facebook support
 * linkedin support
+
+v 0.0.4
+=======
+* facebook credential method 
+
+v 0.0.5
+=======
+* google credential method
+
+v 0.0.6
+=======
+* linkedin credential method